Simplified ppm export functions

+++ Simple CPP Ray Tracer/src/ppm.cpp
@@ -13,7 +13,7 @@ void exportPPMtxt(RGB *data, Size size)
 	printf("P%d\n%d %d\n%d\n", 3, size.width, size.height, 255);
 	for (uint y = 0; y < size.height; y++)
 		for (uint x = 0; x < size.width; x++) {
-			RGB t = data[y * (int)size.height + x];
+			RGB t = data[y * size.height + x];
 			printf("%d %d %d\n", t.R, t.G, t.B);
 		}
 }
@@ -24,7 +24,7 @@ void exportPPMbin(RGB* data, Size size)
 	printf("P%d\r%d %d\r%d\r", 6, size.width, size.height, 255);
 	for (uint y = 0; y < size.height; y++)
 		for (uint x = 0; x <= size.width; x++) {
-			RGB t = data[y * (int)size.height + x];
+			RGB t = data[y * size.height + x];
 			printf("%c%c%c", t.R, t.G, t.B);
 		}
 }
